@@ -9,26 +9,26 @@
     "lint": "ng lint",
     "e2e": "ng e2e",
     "dev:ssr": "ng run rvhl-client:serve-ssr",
-    "serve:ssr": "node dist/rvhl-client/server/main.js",
-    "build:ssr": "ng build --prod && ng run rvhl-client:server:production",
+    "compile:server": "webpack --config webpack.server.config.js --progress --colors",
+    "serve:ssr": "node dist/server",
+    "build:ssr": "npm run build:client-and-server-bundles && npm run compile:server",
+    "build:client-and-server-bundles": "ng build --configuration=en && ng build --configuration=vi  && ng run rvhl-client:server:production  &&  ng run rvhl-client:server:vi &&  ng run rvhl-client:server:en",
     "prerender": "ng run rvhl-client:prerender"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~10.2.0",
-    "@angular/cdk": "^10.2.5",
-    "@angular/common": "~10.2.0",
-    "@angular/compiler": "~10.2.0",
-    "@angular/core": "~10.2.0",
-    "@angular/forms": "~10.2.0",
-    "@angular/material": "^10.2.5",
-    "@angular/platform-browser": "^10.2.0",
-    "@angular/platform-browser-dynamic": "~10.2.0",
-    "@angular/platform-server": "^10.2.0",
-    "@angular/router": "~10.2.0",
-    "@nguniversal/express-engine": "^10.1.0",
-    "@ngx-translate/core": "^13.0.0",
-    "@ngx-translate/http-loader": "^6.0.0",
+    "@angular/animations": "~11.0.2",
+    "@angular/cdk": "^11.0.0",
+    "@angular/common": "~11.0.2",
+    "@angular/compiler": "~11.0.2",
+    "@angular/core": "~11.0.2",
+    "@angular/forms": "~11.0.2",
+    "@angular/material": "^11.0.0",
+    "@angular/platform-browser": "^11.0.2",
+    "@angular/platform-browser-dynamic": "~11.0.2",
+    "@angular/platform-server": "^11.0.2",
+    "@angular/router": "~11.0.2",
+    "@nguniversal/express-engine": "^11.0.0",
     "angular-responsive-carousel": "^1.3.11",
     "express": "^4.15.2",
     "ngx-skeleton-loader": "^2.4.4",
@@ -38,25 +38,26 @@
     "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1002.0",
-    "@angular/cli": "~10.2.0",
-    "@angular/compiler-cli": "~10.2.0",
-    "@nguniversal/builders": "^10.1.0",
+    "@angular-devkit/build-angular": "~0.1100.2",
+    "@angular/cli": "~11.0.2",
+    "@angular/compiler-cli": "~11.0.2",
+    "@angular/localize": "^11.0.2",
+    "@nguniversal/builders": "^11.0.0",
     "@types/express": "^4.17.0",
-    "@types/node": "^12.11.1",
-    "@types/jasmine": "~3.5.0",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
+    "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.5.0",
+    "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "karma": "~5.1.1",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~3.3.0",
+    "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~3.9.5"
+    "typescript": "~4.0.5"
   }
 }
